Physics layer added
Also general cleanup
