+++ src/Physics/index.ts
@@ -1,0 +1,15 @@
+/** Name:   BangJs.Physics.index.ts
+ *  Desc:   Index file for all physics classes
+ *  Author: Jimy Houlbrook
+ *  Date:   28/04/24
+ */
+
+import { Engine } from "./engine";
+import { Layer } from "./layer";
+
+const Physics = {
+    Engine,
+    Layer
+}    
+
+module.exports = Physics;
